+++ server/src/main/java/org/elasticsearch/index/analysis/CustomAnalyzerProvider.java
@@ -24,9 +24,9 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.IndexSettings;
 import org.elasticsearch.index.mapper.TextFieldMapper;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
+
+import static org.elasticsearch.index.analysis.AnalyzerComponents.createComponents;
 
 /**
  * A custom analyzer that is built out of a single {@link org.apache.lucene.analysis.Tokenizer} and a list
@@ -69,99 +69,8 @@ public class CustomAnalyzerProvider extends AbstractIndexAnalyzerProvider<Analyz
         }
     }
 
-    static AnalyzerComponents createComponents(String name, Settings analyzerSettings,
-                                                      final Map<String, TokenizerFactory> tokenizers,
-                                                      final Map<String, CharFilterFactory> charFilters,
-                                                      final Map<String, TokenFilterFactory> tokenFilters) {
-        String tokenizerName = analyzerSettings.get("tokenizer");
-        if (tokenizerName == null) {
-            throw new IllegalArgumentException("Custom Analyzer [" + name + "] must be configured with a tokenizer");
-        }
-
-        TokenizerFactory tokenizer = tokenizers.get(tokenizerName);
-        if (tokenizer == null) {
-            throw new IllegalArgumentException("Custom Analyzer [" + name + "] failed to find tokenizer under name " +
-                "[" + tokenizerName + "]");
-        }
-
-        List<String> charFilterNames = analyzerSettings.getAsList("char_filter");
-        List<CharFilterFactory> charFiltersList = new ArrayList<>(charFilterNames.size());
-        for (String charFilterName : charFilterNames) {
-            CharFilterFactory charFilter = charFilters.get(charFilterName);
-            if (charFilter == null) {
-                throw new IllegalArgumentException("Custom Analyzer [" + name + "] failed to find char_filter under name " +
-                    "[" + charFilterName + "]");
-            }
-            charFiltersList.add(charFilter);
-        }
-
-        int positionIncrementGap = TextFieldMapper.Defaults.POSITION_INCREMENT_GAP;
-
-        positionIncrementGap = analyzerSettings.getAsInt("position_increment_gap", positionIncrementGap);
-
-        int offsetGap = analyzerSettings.getAsInt("offset_gap", -1);
-
-        List<String> tokenFilterNames = analyzerSettings.getAsList("filter");
-        List<TokenFilterFactory> tokenFilterList = new ArrayList<>(tokenFilterNames.size());
-        for (String tokenFilterName : tokenFilterNames) {
-            TokenFilterFactory tokenFilter = tokenFilters.get(tokenFilterName);
-            if (tokenFilter == null) {
-                throw new IllegalArgumentException("Custom Analyzer [" + name + "] failed to find filter under name " +
-                    "[" + tokenFilterName + "]");
-            }
-            tokenFilter = tokenFilter.getChainAwareTokenFilterFactory(tokenizer, charFiltersList, tokenFilterList, tokenFilters::get);
-            tokenFilterList.add(tokenFilter);
-        }
-
-        return new AnalyzerComponents(tokenizerName, tokenizer,
-            charFiltersList.toArray(new CharFilterFactory[charFiltersList.size()]),
-            tokenFilterList.toArray(new TokenFilterFactory[tokenFilterList.size()])
-        );
-    }
-
     @Override
     public Analyzer get() {
         return this.customAnalyzer;
-    }
-
-    public static class AnalyzerComponents {
-        private final String tokenizerName;
-        private final TokenizerFactory tokenizerFactory;
-        private final CharFilterFactory[] charFilters;
-        private final TokenFilterFactory[] tokenFilters;
-        private final AnalysisMode analysisMode;
-
-        AnalyzerComponents(String tokenizerName, TokenizerFactory tokenizerFactory, CharFilterFactory[] charFilters,
-                           TokenFilterFactory[] tokenFilters) {
-            this.tokenizerName = tokenizerName;
-            this.tokenizerFactory = tokenizerFactory;
-            this.charFilters = charFilters;
-            this.tokenFilters = tokenFilters;
-            AnalysisMode mode = AnalysisMode.ALL;
-            for (TokenFilterFactory f : tokenFilters) {
-                mode = mode.merge(f.getAnalysisMode());
-            }
-            this.analysisMode = mode;
-        }
-
-        public String getTokenizerName() {
-            return tokenizerName;
-        }
-
-        public TokenizerFactory getTokenizerFactory() {
-            return tokenizerFactory;
-        }
-
-        public TokenFilterFactory[] getTokenFilters() {
-            return tokenFilters;
-        }
-
-        public CharFilterFactory[] getCharFilters() {
-            return charFilters;
-        }
-
-        public AnalysisMode analysisMode() {
-            return this.analysisMode;
-        }
     }
 }
