+++ server/src/main/java/org/elasticsearch/index/analysis/ReloadableCustomAnalyzer.java
@@ -24,7 +24,6 @@ import org.apache.lucene.analysis.TokenStream;
 import org.apache.lucene.analysis.Tokenizer;
 import org.apache.lucene.util.CloseableThreadLocal;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.index.analysis.CustomAnalyzerProvider.AnalyzerComponents;
 
 import java.io.Reader;
 import java.util.Map;
@@ -124,7 +123,7 @@ public final class ReloadableCustomAnalyzer extends Analyzer implements TokenFil
                                     final Map<String, TokenizerFactory> tokenizers,
                                     final Map<String, CharFilterFactory> charFilters,
                                     final Map<String, TokenFilterFactory> tokenFilters) {
-        AnalyzerComponents components = CustomAnalyzerProvider.createComponents(name, settings, tokenizers, charFilters, tokenFilters);
+        AnalyzerComponents components = AnalyzerComponents.createComponents(name, settings, tokenizers, charFilters, tokenFilters);
         this.components = components;
     }
